finish controller for viewing awaiting approval
